requirements.txt
protobuf==3.20.3
python-dotenv==1.1.1
streamlit==1.14.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pylint>=2.17.0

